+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Display local authorities and their details
 - Allow local authority records to be edited
 - Allow new local authority records to be added
+- Allow local authority records to be deleted
 
 ### Changed